+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [0.1.0] - 17/11/2025
+
+### Added
+- Delete supply batch
+- Added filter and order products
+
 ## [0.1.0] - 16/11/2025
 
 ### Added